Fix tests for Transformer reimplementation
