+++ requirements.txt
@@ -16,7 +16,7 @@ requests==2.20.0
 six==1.11.0
 simplejson==3.16.0
 setuptools>=38.6.0
-SPARQLTransformer==1.6.10
+SPARQLTransformer==1.7.0
 SPARQLWrapper==1.8.2
 Werkzeug==0.14.1
 PyGithub==1.43.5
